move request definitions in its own module
